CSIntValueStoreProperty: Support lambda for default value
This change allows the default value for `CSIntValueStoreProperty` and `CSIntValuePresetProperty` to be provided by a lambda function.

This is useful in scenarios where the default value might depend on other factors or needs to be calculated dynamically.

New constructor and property extension functions have been added to support this, and existing contexts (`PresetStoreContext`, `CustomStoreContext`) have been updated accordingly.