Make default value of CSStringValuePresetProperty lazy
This change allows for the default value of a `CSStringValuePresetProperty` to be computed lazily by accepting a function that returns the default string.

A new constructor and a corresponding extension function `CSPreset<*, *>.property` have been added to support this functionality. The existing constructor and extension function that accept a static string default value remain for backward compatibility.